+++ packages/sdks/snippets/angular-ssr/src/app/announcement-bar/announcement-bar.component.ts
@@ -4,15 +4,22 @@
  * src/app/announcement-bar/announcement-bar.component.ts
  */
 
+import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { type BuilderContent } from '@builder.io/sdk-angular';
+import { Content, type BuilderContent } from '@builder.io/sdk-angular';
 
 @Component({
   selector: 'app-announcement-bar',
+  standalone: true,
+  imports: [Content, CommonModule],
   template: `
     <ng-container *ngIf="content">
-      <content [model]="model" [content]="content" [apiKey]="apiKey"></content>
+      <builder-content
+        [model]="model"
+        [content]="content"
+        [apiKey]="apiKey"
+      ></builder-content>
     </ng-container>
 
     <!-- Your content coming from your app (or also Builder) -->