+++ packages/sdks/snippets/angular/src/app/announcement-bar/announcement-bar.component.ts
@@ -4,14 +4,25 @@
  * src/app/announcement-bar/announcement-bar.component.ts
  */
 
+import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
-import { fetchOneEntry, type BuilderContent } from '@builder.io/sdk-angular';
+import {
+  Content,
+  fetchOneEntry,
+  type BuilderContent,
+} from '@builder.io/sdk-angular';
 
 @Component({
   selector: 'app-announcement-bar',
+  standalone: true,
+  imports: [CommonModule, Content],
   template: `
     <ng-container *ngIf="content">
-      <content [model]="model" [content]="content" [apiKey]="apiKey"></content>
+      <builder-content
+        [model]="model"
+        [content]="content"
+        [apiKey]="apiKey"
+      ></builder-content>
     </ng-container>
 
     <!-- Your content coming from your app (or also Builder) -->