+++ packages/sdks/snippets/angular/src/app/app.module.ts
@@ -6,19 +6,24 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { RouterModule } from '@angular/router';
-import { Content } from '@builder.io/sdk-angular';
 import { AnnouncementBarComponent } from './announcement-bar/announcement-bar.component';
 import { AppComponent } from './app.component';
+import { BlogArticleComponent } from './blog-article/blog-article.component';
 import { CatchAllComponent } from './catch-all/catch-all.component';
+import { ProductEditorialComponent } from './product-editorial/product-editorial.component';
 
 @NgModule({
-  declarations: [AppComponent, AnnouncementBarComponent, CatchAllComponent],
-  // add Content to imports
+  declarations: [AppComponent],
   imports: [
     BrowserModule,
-    Content,
+    BlogArticleComponent,
+    ProductEditorialComponent,
+    AnnouncementBarComponent,
+    CatchAllComponent,
     RouterModule.forRoot([
       { path: 'announcements/:id', component: AnnouncementBarComponent },
+      { path: 'blogs/new-product-line', component: BlogArticleComponent },
+      { path: 'products/:id', component: ProductEditorialComponent },
       { path: '**', component: CatchAllComponent },
     ]),
   ],